Added basic post capabilities to the entry page.
Currently the expirements will post with the two fields specified.
The Saga that is built can be used for any of the backend models.
Window alerts are currently being used as the messaging medium this
should be changed to a in app solution in the future.

Response recieved a few updates to make use of status text from our
API standard response error messages.